[FIX] We must now inherit from the base model, instead of monkey-patching
