+++ base_phone/models/__init__.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 
+from . import base
 from . import res_company
 from . import res_partner
 from . import phone_common
